Refactor base classes
