+++ ns_pipelines/base/__init__.py
@@ -1,0 +1,8 @@
+from .api import APIPromptExtractor
+from .base import IndependentPipeline, DependentPipeline
+
+__all__ = [
+    "APIPromptExtractor",
+    "IndependentPipeline",
+    "DependentPipeline",
+]
